Create module for database utilities
