+++ gramps/cli/arghandler.py
@@ -46,7 +46,7 @@ import re
 from gramps.gen.recentfiles import recent_files
 from gramps.gen.utils.file import rm_tempdir, get_empty_tempdir
 from .clidbman import CLIDbManager, NAME_FILE, find_locker_name
-
+from gramps.gen.db.utils import make_database
 from gramps.gen.plug import BasePluginManager
 from gramps.gen.plug.report import CATEGORY_BOOK, CATEGORY_CODE, BookList
 from .plug import cl_report, cl_book
@@ -501,7 +501,7 @@ class ArgHandler:
                 else:
                     self.imp_db_path = get_empty_tempdir("import_dbdir")
                     dbid = config.get('database.backend')
-                    newdb = self.dbstate.make_database(dbid)
+                    newdb = make_database(dbid)
                     newdb.write_version(self.imp_db_path)
 
                 try:
