+++ gramps/gen/filters/rules/test/person_rules_test.py
@@ -24,7 +24,7 @@ Unittest that tests person-specific filter rules
 import unittest
 import os
 
-from gramps.gen.merge.diff import import_as_dict
+from gramps.gen.db.utils import import_as_dict
 from gramps.cli.user import User
 from gramps.gen.filters import GenericFilter
 from gramps.gen.const import DATA_DIR
