+++ gramps/gen/lib/test/struct_test.py
@@ -25,7 +25,7 @@ import os
 
 from  .. import (Person, Family, Event, Source, Place, Citation,
                  Repository, Media, Note, Tag)
-from gramps.gen.merge.diff import import_as_dict
+from gramps.gen.db.utils import import_as_dict
 from gramps.cli.user import User
 from gramps.gen.const import DATA_DIR
